pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-client"
version = "0.1.0"
description = "Ein universeller LLM-Client für OpenAI, Groq und Ollama."
authors = [
  { name = "Daniel Gaida", email = "daniel.gaida@th-koeln.de" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0.1",
  "openai>=1.51.0",
  "groq>=0.5.0",
  "ollama>=0.1.9"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0"
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-v --tb=short"
testpaths = ["tests"]

[tool.ruff]
select = ["E", "F", "I"]
line-length = 100
